separate gzstream test
