+++ Code/JavaWrappers/gmwrapper/src-test/org/RDKit/SuppliersTests.java
@@ -167,28 +167,6 @@ public class SuppliersTests extends GraphMolTest {
                 if((i%1000)==0) System.err.printf("Done: %s\n",i);
             }
 	}
-  
-        @Test
-        public void test11GZstream() {
-          // NCI_aids_few.sdf.gz
-          File base = getRdBase();
-          File gzpath  = new File(base, "Code" + File.separator + "GraphMol" + File.separator +
-                                  "FileParsers" + File.separator + "test_data");
-          File fileN = new File(gzpath, "NCI_aids_few.sdf.gz");
-          assertTrue(fileN.exists());
-          gzstream stream = new gzstream(fileN.getPath());
-          ForwardSDMolSupplier suppl = new ForwardSDMolSupplier(stream);
-          assertFalse(suppl.atEnd());
-          ArrayList<ROMol> ms = new ArrayList<ROMol>();
-          ROMol m;
-          do {
-            m = suppl.next();
-            if (m != null)
-              ms.add(m);
-          } while (!suppl.atEnd());
-          assertEquals(16, ms.size());
-
-        }
 
 	public static void main(String args[]) {
 		org.junit.runner.JUnitCore.main("org.RDKit.SuppliersTests");
